[#2425] - Nombra los e2e de lectura por la ruta que ejercitan
Los dos specs de indexado de obra estaban al revés del precedente de colección:
el que se llamaba como el recurso cubría el catálogo, y el detalle vivía bajo el
nombre de la ruta vieja. El intercambio va en dos tiempos porque el nombre de
destino estaba ocupado: primero el catálogo pasa a su forma plural, y recién
entonces el detalle toma el singular.

Los tres specs de comportamiento de la página y la utilidad de fixtures
acompañan, y con ellos los nombres de caso, que anunciaban una ruta que ya no
existe.

Incluye una aserción de la home que esperaba una URL de la ruta vieja: con el
renombre habría roto el gate e2e.
