+++ e2e/seo/literary-work.spec.ts
@@ -1,82 +1,100 @@
 /**
- * Tests e2e de SEO para el catálogo de obras (`/literary-work`).
+ * Tests e2e de SEO para la página de lectura (`/literary-work/:slug`, entidad LiteraryWork).
  *
- * La tanda completa corre sin `fixme`: la página no difiere nada, así que el HTML servido trae la
- * tabla entera con el enlace a la lectura de cada obra.
+ * Sobre el HTML server-rendered (lo que ve el crawler, sin ejecutar JS):
+ *  - A. Status 404 real del SSR para una obra inexistente (no 200 con página vacía).
+ *  - B. Meta tags: <title> con la obra, description, og:/twitter: title y description,
+ *       canonical a la URL de la obra, robots indexable, keywords y author (señal E-E-A-T).
+ *  - C. Datos estructurados: JSON-LD Article (headline, datePublished, author Person[],
+ *       publisher Organization; LiteraryWork no expone updatedAt y el builder omite dateModified)
+ *       y BreadcrumbList.
+ *  - D. Bloques sitewide Organization y WebSite.
+ *  - E. La tanda completa de invariantes de una página indexable, H1 único y cuerpo saneado.
+ *
+ * El contenido de prueba lo cura el equipo en los datasets (development local / staging CI).
  */
 import { test, expect } from '@playwright/test';
+import type { Article, BreadcrumbList, WithContext } from 'schema-dts';
 
 import { parseJsonLdBlocks, getMetaContent, getTitleText, getCanonicalHref } from '../_utils/seo';
 import { assertValidJsonLd } from '@testing/json-ld-validation';
 import { collectIndexableHtmlViolations } from '../_utils/seo-invariants';
-import { SCHEMA_IDS, SITEWIDE_SCHEMA_IDS } from '../_utils/seo-fixtures';
+import { STABLE_SLUGS, SCHEMA_IDS, SITEWIDE_SCHEMA_IDS } from '../_utils/seo-fixtures';
 
-const catalogPath = '/literary-work';
-const requiredJsonLdIds = [
-	...SITEWIDE_SCHEMA_IDS,
-	SCHEMA_IDS.literaryWorkCatalog,
-	SCHEMA_IDS.breadcrumbLiteraryWorkCatalog,
-];
+const literaryWorkPath = `/literary-work/${STABLE_SLUGS.literaryWork}`;
+const requiredJsonLdIds = [...SITEWIDE_SCHEMA_IDS, SCHEMA_IDS.article, SCHEMA_IDS.breadcrumbLiteraryWork];
 
-test.describe('literary-work — HTML server-rendered del catálogo', () => {
+test('literary-work — A: una obra inexistente responde 404 real en SSR', async ({ request }) => {
+	const response = await request.get('/literary-work/obra-inexistente-e2e');
+	expect(response.status()).toBe(404);
+});
+
+test.describe('literary-work — HTML server-rendered de una obra existente', () => {
 	let html: string;
 
 	test.beforeAll(async ({ request }) => {
-		const response = await request.get(catalogPath);
-		expect(response.status(), 'el catálogo de obras no respondió 200').toBe(200);
+		const response = await request.get(literaryWorkPath);
+		// Sin esto, un slug que desapareció del dataset hace fallar los cinco casos por "expected truthy",
+		// sin nombrar la causa. Se afirma en vez de saltearse: es el skip el que ya dejó un verde engañoso.
+		expect(response.status(), `No existe literaryWork con slug "${STABLE_SLUGS.literaryWork}" en el dataset`).toBe(200);
 		html = await response.text();
 	});
 
-	test('A: meta tags en el HTML server-rendered', () => {
-		expect(getTitleText(html)).toContain('Obras');
+	test('B: meta tags en el HTML server-rendered', () => {
+		expect(getTitleText(html)).toBeTruthy();
 		expect(getMetaContent(html, 'description')).toBeTruthy();
 		expect(getMetaContent(html, 'og:title')).toBeTruthy();
 		expect(getMetaContent(html, 'og:description')).toBeTruthy();
 		expect(getMetaContent(html, 'twitter:title')).toBeTruthy();
 		expect(getMetaContent(html, 'twitter:description')).toBeTruthy();
-		expect(getCanonicalHref(html)).toContain(catalogPath);
+		expect(getCanonicalHref(html)).toContain(literaryWorkPath);
 		expect(getMetaContent(html, 'keywords')).toBeTruthy();
-		// La ausencia de `noindex` y no la presencia de `index`, que pasaría con las dos políticas.
+		expect(getMetaContent(html, 'author')).toBeTruthy();
+		// Se afirma la ausencia de `noindex` y no la presencia de `index`, que pasaría con las dos
+		// políticas: `noindex` contiene ese substring.
 		expect(getMetaContent(html, 'robots')).not.toContain('noindex');
 	});
 
-	test('B: JSON-LD CollectionPage con el listado del catálogo', async () => {
-		const catalog = parseJsonLdBlocks(html).get(SCHEMA_IDS.literaryWorkCatalog);
-		await assertValidJsonLd(catalog);
+	test('C: JSON-LD Article de la obra', async () => {
+		const article = parseJsonLdBlocks(html).get(SCHEMA_IDS.article) as WithContext<Article> | undefined;
+		await assertValidJsonLd(article);
 
-		const mainEntity = catalog?.['mainEntity'] as Record<string, unknown>;
-		expect(mainEntity?.['@type']).toBe('ItemList');
-		expect(Number(mainEntity?.['numberOfItems'])).toBeGreaterThan(0);
-
-		// El catálogo vive en `/literary-work` y cada obra en `/literary-work/<slug>`: el elemento apunta a la
-		// lectura, no al listado.
-		const [firstItem] = (mainEntity?.['itemListElement'] ?? []) as Record<string, unknown>[];
-		expect(firstItem?.['position']).toBe(1);
-		expect(String(firstItem?.['url'])).toContain('/literary-work/');
+		expect(article?.['@type']).toBe('Article');
+		expect(article?.headline).toBeTruthy();
+		expect(article?.datePublished).toBeTruthy();
+		expect(String(article?.mainEntityOfPage)).toContain('/literary-work/');
 	});
 
-	test('B: JSON-LD BreadcrumbList', async () => {
-		const breadcrumb = parseJsonLdBlocks(html).get(SCHEMA_IDS.breadcrumbLiteraryWorkCatalog);
+	test('C: JSON-LD BreadcrumbList', async () => {
+		const breadcrumb = parseJsonLdBlocks(html).get(SCHEMA_IDS.breadcrumbLiteraryWork) as
+			WithContext<BreadcrumbList> | undefined;
 		await assertValidJsonLd(breadcrumb);
 
-		expect((breadcrumb?.['itemListElement'] as unknown[])?.length).toBe(2);
+		const items = breadcrumb?.itemListElement;
+		expect(Array.isArray(items) ? items.length : 0).toBeGreaterThanOrEqual(2);
 	});
 
-	test('C: bloques sitewide Organization y WebSite presentes', () => {
+	test('D: bloques sitewide Organization y WebSite presentes', () => {
 		const blocks = parseJsonLdBlocks(html);
 
 		expect(blocks.get(SCHEMA_IDS.organization)?.['@type']).toBe('Organization');
 		expect(blocks.get(SCHEMA_IDS.website)?.['@type']).toBe('WebSite');
 	});
 
-	test('D: invariantes de página indexable (ssr, title, canonical, robots, h1, cuerpo, enlaces, jsonld)', async () => {
+	test('E: invariantes de página indexable (ssr, title, canonical, robots, h1, cuerpo, enlaces, jsonld)', async () => {
 		expect(
 			await collectIndexableHtmlViolations(html, {
-				path: catalogPath,
+				path: literaryWorkPath,
 				requiredJsonLdIds,
-				requiredInternalLinkPrefix: '/literary-work/',
-				h1Pattern: /\d+ Obras?/,
+				requiredInternalLinkPrefix: '/author/',
 			}),
 		).toEqual([]);
+	});
+
+	test('E: H1 único con contenido real y cuerpo saneado', () => {
+		expect(html.match(/<h1[^>]*>/g) ?? []).toHaveLength(1);
+		expect(html).toContain('<article');
+		// El markdown crudo no cruza al frontend: sin ** literales dentro del artículo.
+		expect(html).not.toMatch(/<article[^>]*>[\s\S]*\*\*[\s\S]*<\/article>/);
 	});
 });
