+++ src/api/_helpers/cache-control.ts
@@ -16,7 +16,7 @@ const READ_CACHE_STALE_WHILE_REVALIDATE = 604800; // 7 días
 const BROWSER_CACHE_CONTROL = 'public, max-age=0, must-revalidate';
 
 /**
- * Decide si `/read` es cacheable en el entorno actual. Fuera de producción no lo es (coherente con
+ * Decide si la página de lectura es cacheable en el entorno actual. Fuera de producción no lo es (coherente con
  * `noindexNonProduction`): un preview comparte el CDN y serviría contenido de un dataset que no es
  * el público. Exportada para que el middleware corte antes de bufferizar el body, sin duplicar la
  * condición: la política sigue teniendo un solo dueño.
@@ -26,7 +26,7 @@ export function isReadCacheEnabled(): boolean {
 }
 
 /**
- * Aplica a una respuesta cacheable de `/read` los headers de caché de borde: `s-maxage` corto
+ * Aplica a una respuesta cacheable de la página de lectura los headers de caché de borde: `s-maxage` corto
  * (interruptor `environment.readCacheSMaxAge`) con `stale-while-revalidate` largo, efectivos solo
  * en el CDN de Vercel, más un `Cache-Control` que mantiene fresco al browser.
  *
